Split the code around `Draggable` and `DragTarget`
to improve developer happiness and lower cognitive load while working on parts of it.
