+++ packages/flutter/lib/src/widgets/draggable/drag_avatar.dart
@@ -1,0 +1,195 @@
+import 'package:flutter/gestures.dart';
+import 'package:flutter/rendering.dart';
+
+import 'package:flutter/src/widgets/basic.dart';
+import 'package:flutter/src/widgets/binding.dart';
+import 'package:flutter/src/widgets/framework.dart';
+import 'package:flutter/src/widgets/overlay.dart';
+
+import 'drag_target_state.dart';
+import 'drag_update_callback.dart';
+
+enum _DragEndKind { dropped, canceled }
+typedef _OnDragEnd = void Function(Velocity velocity, Offset offset, bool wasAccepted);
+
+// The lifetime of this object is a little dubious right now. Specifically, it
+// lives as long as the pointer is down. Arguably it should self-immolate if the
+// overlay goes away. _DraggableState has some delicate logic to continue
+// needing this object pointer events even after it has been disposed.
+class DragAvatar<T extends Object> extends Drag {
+  DragAvatar({
+    required this.overlayState,
+    this.data,
+    this.axis,
+    required Offset initialPosition,
+    this.dragStartPoint = Offset.zero,
+    this.feedback,
+    this.feedbackOffset = Offset.zero,
+    this.onDragUpdate,
+    this.onDragEnd,
+    required this.ignoringFeedbackSemantics,
+  }) : assert(overlayState != null),
+        assert(ignoringFeedbackSemantics != null),
+        assert(dragStartPoint != null),
+        assert(feedbackOffset != null),
+        _position = initialPosition {
+    _entry = OverlayEntry(builder: _build);
+    overlayState.insert(_entry!);
+    updateDrag(initialPosition);
+  }
+
+  final T? data;
+  final Axis? axis;
+  final Offset dragStartPoint;
+  final Widget? feedback;
+  final Offset feedbackOffset;
+  final DragUpdateCallback? onDragUpdate;
+  final _OnDragEnd? onDragEnd;
+  final OverlayState overlayState;
+  final bool ignoringFeedbackSemantics;
+
+  DragTargetState<Object>? _activeTarget;
+  final List<DragTargetState<Object>> _enteredTargets = <DragTargetState<Object>>[];
+  Offset _position;
+  Offset? _lastOffset;
+  OverlayEntry? _entry;
+
+  @override
+  void update(DragUpdateDetails details) {
+    final Offset oldPosition = _position;
+    _position += _restrictAxis(details.delta);
+    updateDrag(_position);
+    if (onDragUpdate != null && _position != oldPosition) {
+      onDragUpdate!(details);
+    }
+  }
+
+  @override
+  void end(DragEndDetails details) {
+    finishDrag(_DragEndKind.dropped, _restrictVelocityAxis(details.velocity));
+  }
+
+
+  @override
+  void cancel() {
+    finishDrag(_DragEndKind.canceled);
+  }
+
+  void updateDrag(Offset globalPosition) {
+    _lastOffset = globalPosition - dragStartPoint;
+    _entry!.markNeedsBuild();
+    final HitTestResult result = HitTestResult();
+    WidgetsBinding.instance!.hitTest(result, globalPosition + feedbackOffset);
+
+    final List<DragTargetState<Object>> targets = _getDragTargets(result.path).toList();
+
+    bool listsMatch = false;
+    if (targets.length >= _enteredTargets.length && _enteredTargets.isNotEmpty) {
+      listsMatch = true;
+      final Iterator<DragTargetState<Object>> iterator = targets.iterator;
+      for (int i = 0; i < _enteredTargets.length; i += 1) {
+        iterator.moveNext();
+        if (iterator.current != _enteredTargets[i]) {
+          listsMatch = false;
+          break;
+        }
+      }
+    }
+
+    // If everything's the same, report moves, and bail early.
+    if (listsMatch) {
+      for (final DragTargetState<Object> target in _enteredTargets) {
+        target.didMove(this);
+      }
+      return;
+    }
+
+    // Leave old targets.
+    _leaveAllEntered();
+
+    // Enter new targets.
+    final DragTargetState<Object>? newTarget = targets.cast<DragTargetState<Object>?>().firstWhere(
+          (DragTargetState<Object>? target) {
+        if (target == null)
+          return false;
+        _enteredTargets.add(target);
+        return target.didEnter(this);
+      },
+      orElse: () => null,
+    );
+
+    // Report moves to the targets.
+    for (final DragTargetState<Object> target in _enteredTargets) {
+      target.didMove(this);
+    }
+
+    _activeTarget = newTarget;
+  }
+
+  Iterable<DragTargetState<Object>> _getDragTargets(Iterable<HitTestEntry> path) sync* {
+    // Look for the RenderBoxes that corresponds to the hit target (the hit target
+    // widgets build RenderMetaData boxes for us for this purpose).
+    for (final HitTestEntry entry in path) {
+      final HitTestTarget target = entry.target;
+      if (target is RenderMetaData) {
+        final dynamic metaData = target.metaData;
+        if (metaData is DragTargetState && metaData.isExpectedDataType(data, T))
+          yield metaData;
+      }
+    }
+  }
+
+  void _leaveAllEntered() {
+    for (int i = 0; i < _enteredTargets.length; i += 1)
+      _enteredTargets[i].didLeave(this);
+    _enteredTargets.clear();
+  }
+
+  void finishDrag(_DragEndKind endKind, [ Velocity? velocity ]) {
+    bool wasAccepted = false;
+    if (endKind == _DragEndKind.dropped && _activeTarget != null) {
+      _activeTarget!.didDrop(this);
+      wasAccepted = true;
+      _enteredTargets.remove(_activeTarget);
+    }
+    _leaveAllEntered();
+    _activeTarget = null;
+    _entry!.remove();
+    _entry = null;
+    // TODO(ianh): consider passing _entry as well so the client can perform an animation.
+    if (onDragEnd != null)
+      onDragEnd!(velocity ?? Velocity.zero, _lastOffset!, wasAccepted);
+  }
+
+  Widget _build(BuildContext context) {
+    final RenderBox box = overlayState.context.findRenderObject()! as RenderBox;
+    final Offset overlayTopLeft = box.localToGlobal(Offset.zero);
+    return Positioned(
+      left: _lastOffset!.dx - overlayTopLeft.dx,
+      top: _lastOffset!.dy - overlayTopLeft.dy,
+      child: IgnorePointer(
+        child: feedback,
+        ignoringSemantics: ignoringFeedbackSemantics,
+      ),
+    );
+  }
+
+  Velocity _restrictVelocityAxis(Velocity velocity) {
+    if (axis == null) {
+      return velocity;
+    }
+    return Velocity(
+      pixelsPerSecond: _restrictAxis(velocity.pixelsPerSecond),
+    );
+  }
+
+  Offset _restrictAxis(Offset offset) {
+    if (axis == null) {
+      return offset;
+    }
+    if (axis == Axis.horizontal) {
+      return Offset(offset.dx, 0.0);
+    }
+    return Offset(0.0, offset.dy);
+  }
+}
