+++ packages/flutter/lib/src/widgets/draggable/drag_update_callback.dart
@@ -1,0 +1,7 @@
+import 'package:flutter/gestures.dart';
+
+/// Signature for when a [Draggable] is dragged across the screen.
+///
+/// Used by [Draggable.onDragUpdate].
+typedef DragUpdateCallback = void Function(DragUpdateDetails details);
+
