+++ packages/flutter/lib/widgets.dart
@@ -36,7 +36,7 @@ export 'src/widgets/debug.dart';
 export 'src/widgets/desktop_text_selection_toolbar_layout_delegate.dart';
 export 'src/widgets/dismissible.dart';
 export 'src/widgets/disposable_build_context.dart';
-export 'src/widgets/drag_target.dart';
+export 'src/widgets/draggable/drag_target.dart';
 export 'src/widgets/draggable_scrollable_sheet.dart';
 export 'src/widgets/dual_transition_builder.dart';
 export 'src/widgets/editable_text.dart';
